requirements.txt
Django==4.2.7
djangorestframework==3.14.0
web3==6.11.1
psycopg2-binary==2.9.9
django-redis==5.4.0
gunicorn==21.2.0
python-dotenv==1.0.0